Upgrade VisItBridge to visit2.7.0

+++ Library/VisItLib/avt/DBAtts/MetaData/avtSimulationInformation.C
@@ -1,8 +1,8 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2010, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
-* LLNL-CODE-400124
+* LLNL-CODE-442911
 * All rights reserved.
 *
 * This file is  part of VisIt. For  details, see https://visit.llnl.gov/.  The
@@ -816,7 +816,7 @@ avtSimulationInformation::RemoveGenericCommands(int index)
 int
 avtSimulationInformation::GetNumGenericCommands() const
 {
-    return genericCommands.size();
+    return (int)genericCommands.size();
 }
 
 // ****************************************************************************
@@ -967,7 +967,7 @@ avtSimulationInformation::RemoveCustomCommands(int index)
 int
 avtSimulationInformation::GetNumCustomCommands() const
 {
-    return customCommands.size();
+    return (int)customCommands.size();
 }
 
 // ****************************************************************************
