+++ Library/VisItLib/avt/DBAtts/SIL/avtSIL.h
@@ -1,8 +1,8 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2010, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
-* LLNL-CODE-400124
+* LLNL-CODE-442911
 * All rights reserved.
 *
 * This file is  part of VisIt. For  details, see https://visit.llnl.gov/.  The
@@ -181,7 +181,7 @@ class DBATTS_API avtSIL
 
     int                               AddSet(avtSILSet_p);
     int                               GetNumRealSets() const
-                                           { return sets.size(); };
+                                           { return static_cast<int>(sets.size()); };
     EntryType                         GetCollectionSource(int index, 
                                                           avtSILArray_p  &outArray, 
                                                           avtSILMatrix_p &outMatrix, int &outIndex);
