+++ Library/VisItLib/avt/DBAtts/SIL/avtSILEnumeratedNamespace.h
@@ -1,8 +1,8 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2010, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
-* LLNL-CODE-400124
+* LLNL-CODE-442911
 * All rights reserved.
 *
 * This file is  part of VisIt. For  details, see https://visit.llnl.gov/.  The
@@ -76,7 +76,7 @@ class DBATTS_API avtSILEnumeratedNamespace : public avtSILNamespace
     virtual            ~avtSILEnumeratedNamespace() {;};
 
     virtual int                      GetNumberOfElements(void) const
-                                                 { return elements.size(); };
+                                                 { return static_cast<int>(elements.size()); };
     virtual int                      GetElement(int idx) const
                                                  { return elements[idx]; };
     virtual NamespaceAttributes     *GetAttributes(void) const;
