+++ Library/VisItLib/avt/Pipeline/Data/avtAverageValueRF.C
@@ -1,8 +1,8 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2010, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
-* LLNL-CODE-400124
+* LLNL-CODE-442911
 * All rights reserved.
 *
 * This file is  part of VisIt. For  details, see https://visit.llnl.gov/.  The
@@ -127,7 +127,7 @@ avtAverageValueRF::~avtAverageValueRF()
 
 void
 avtAverageValueRF::GetRayValue(const avtRay *ray, 
-                               unsigned char rgb[3], float depth)
+                               unsigned char rgb[3], double depth)
 {
     //
     // Some compilers do very poor optimizations, so make sure that we don't
@@ -138,7 +138,7 @@ avtAverageValueRF::GetRayValue(const avtRay *ray,
     const bool   *validSample = ray->validSample;
 
     // Only handle one variable, even if there is more.
-    const float  *sample      = ray->sample[0];
+    const double  *sample      = ray->sample[0];
 
     int depthIndex = IndexOfDepth(depth, numSamples);
