+++ Library/VisItLib/avt/Pipeline/Data/avtDistanceToValueRF.h
@@ -1,8 +1,8 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2010, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
-* LLNL-CODE-400124
+* LLNL-CODE-442911
 * All rights reserved.
 *
 * This file is  part of VisIt. For  details, see https://visit.llnl.gov/.  The
@@ -79,7 +79,7 @@ class PIPELINE_API avtDistanceToValueRF : public avtRayFunction
     virtual         ~avtDistanceToValueRF();
 
     virtual void     GetRayValue(const avtRay *,
-                                 unsigned char rgb[3], float);
+                                 unsigned char rgb[3], double);
 
   protected:
     double                  criticalValue;
