+++ Library/VisItLib/avt/Pipeline/Data/avtOpacityMapSamplePointArbitrator.h
@@ -1,8 +1,8 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2010, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
-* LLNL-CODE-400124
+* LLNL-CODE-442911
 * All rights reserved.
 *
 * This file is  part of VisIt. For  details, see https://visit.llnl.gov/.  The
@@ -71,7 +71,7 @@ class PIPELINE_API avtOpacityMapSamplePointArbitrator
                       avtOpacityMapSamplePointArbitrator(avtOpacityMap *, int);
     virtual          ~avtOpacityMapSamplePointArbitrator();
 
-    virtual bool      ShouldOverwrite(float, float);
+    virtual bool      ShouldOverwrite(double, double);
 
   protected:
     avtOpacityMap    *omap;
