+++ Library/VisItLib/avt/Pipeline/Data/avtRayFunction.C
@@ -1,8 +1,8 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2010, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
-* LLNL-CODE-400124
+* LLNL-CODE-442911
 * All rights reserved.
 *
 * This file is  part of VisIt. For  details, see https://visit.llnl.gov/.  The
@@ -113,7 +113,7 @@ avtRayFunction::~avtRayFunction()
 // ****************************************************************************
 
 bool
-avtRayFunction::CanContributeToPicture(int,const float (*)[AVT_VARIABLE_LIMIT])
+avtRayFunction::CanContributeToPicture(int,const double (*)[AVT_VARIABLE_LIMIT])
 {
     return true;
 }
