+++ Library/VisItLib/avt/Pipeline/Data/avtRelativeValueSamplePointArbitrator.C
@@ -1,8 +1,8 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2010, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
-* LLNL-CODE-400124
+* LLNL-CODE-442911
 * All rights reserved.
 *
 * This file is  part of VisIt. For  details, see https://visit.llnl.gov/.  The
@@ -101,8 +101,8 @@ avtRelativeValueSamplePointArbitrator::~avtRelativeValueSamplePointArbitrator()
 // ****************************************************************************
 
 bool
-avtRelativeValueSamplePointArbitrator::ShouldOverwrite(float whatIgot,
-                                                       float whatIcouldHave)
+avtRelativeValueSamplePointArbitrator::ShouldOverwrite(double whatIgot,
+                                                       double whatIcouldHave)
 {
     if (lessThan)
         return (whatIgot < whatIcouldHave);
