+++ Library/VisItLib/avt/Pipeline/Data/avtRelativeValueSamplePointArbitrator.h
@@ -1,8 +1,8 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2010, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2013, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
-* LLNL-CODE-400124
+* LLNL-CODE-442911
 * All rights reserved.
 *
 * This file is  part of VisIt. For  details, see https://visit.llnl.gov/.  The
@@ -70,7 +70,7 @@ class PIPELINE_API avtRelativeValueSamplePointArbitrator
                       avtRelativeValueSamplePointArbitrator(bool, int);
     virtual          ~avtRelativeValueSamplePointArbitrator();
 
-    virtual bool      ShouldOverwrite(float, float);
+    virtual bool      ShouldOverwrite(double, double);
 
   protected:
     bool              lessThan;
